Tests/LibWeb/Text/expected/css-has-invalidation/multiple-has-rules.txt

[add .red (only :has(.red) flips)]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 2
  hasMatchInvocations: 2
  hasResultCacheHits: 0
  hasResultCacheMisses: 2
  styleInvalidations: 1
[add .blue (only :has(.blue) flips, .red stable)]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 2
  hasMatchInvocations: 2
  hasResultCacheHits: 0
  hasResultCacheMisses: 2
  styleInvalidations: 1
